Fix nnue, don't use lclamp

+++ nnue/batch.py
@@ -22,12 +22,10 @@ class batch(ctypes.Structure):
         val1 = torch.ones(self.ind_active).pin_memory().to(device = device, non_blocking = True)
         val2 = torch.ones(self.ind_active).pin_memory().to(device = device, non_blocking = True)
 
-        ind1 = torch.transpose(torch.from_numpy(numpy.ctypeslib.as_array(self.ind1, shape = (self.ind_active, 2))), 0, 1).long().pin_memory().to(device = device, non_blocking = True)
-
-        ind2 = torch.transpose(torch.from_numpy(numpy.ctypeslib.as_array(self.ind2, shape = (self.ind_active, 2))), 0, 1).long().pin_memory().to(device = device, non_blocking = True)
+        ind1 = torch.transpose(torch.from_numpy(numpy.ctypeslib.as_array(self.ind1, shape = (self.ind_active, 2))), 0, 1).pin_memory().to(device = device, non_blocking = True)
+        ind2 = torch.transpose(torch.from_numpy(numpy.ctypeslib.as_array(self.ind2, shape = (self.ind_active, 2))), 0, 1).pin_memory().to(device = device, non_blocking = True)
 
         f1 = torch._sparse_coo_tensor_unsafe(ind1, val1, (self.actual_size, model.FT_IN_DIMS + model.VIRTUAL)).to(device = device, non_blocking = True)
-
         f2 = torch._sparse_coo_tensor_unsafe(ind2, val2, (self.actual_size, model.FT_IN_DIMS + model.VIRTUAL)).to(device = device, non_blocking = True)
 
         f1._coalesced_(True)
